Url-summary/requirements.txt
streamlit
langchain
google-search-results
validators
langchain-google-genai
unstructured
youtube-transcript-api

pytube